functions/thumbnail/requirements.txt
Pillow==8.3.2
google-cloud-storage==2.4.0
